Tags
Tag pages are available via:

    /tags

as well as:

    /tags/some-tag

However if a post is visited, tags that apply to that post are not visible.
Add "tags" partial file, use that partial on post pages, and add a link to
"/tags" on the sidebar. That should cover all the current gaps with regard to
interlinking.